M05_LangChainExpressionLanguage/pyproject.toml
[project]
name = "m05-langchainexpressionlanguage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastapi>=0.117.1",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-google-genai>=2.1.12",
    "langchain-groq>=0.3.8",
    "langserve>=0.3.2",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "sse-starlette>=3.0.2",
    "uvicorn>=0.36.0",
]
